Split replay folder and file management into their own classes
