+++ src/main/java/com/replaymod/core/ReplayMod.java
@@ -1,10 +1,10 @@
 package com.replaymod.core;
 
-import com.google.common.net.PercentEscaper;
 import com.replaymod.compat.ReplayModCompat;
+import com.replaymod.core.files.ReplayFilesService;
+import com.replaymod.core.files.ReplayFoldersService;
 import com.replaymod.core.gui.GuiBackgroundProcesses;
 import com.replaymod.core.gui.GuiReplaySettings;
-import com.replaymod.core.gui.RestoreReplayGui;
 import com.replaymod.core.versions.MCVer;
 import com.replaymod.core.versions.scheduler.Scheduler;
 import com.replaymod.core.versions.scheduler.SchedulerImpl;
@@ -14,12 +14,9 @@ import com.replaymod.recording.ReplayModRecording;
 import com.replaymod.render.ReplayModRender;
 import com.replaymod.replay.ReplayModReplay;
 import com.replaymod.replaystudio.lib.viaversion.api.protocol.version.ProtocolVersion;
-import com.replaymod.replaystudio.replay.ReplayFile;
-import com.replaymod.replaystudio.replay.ZipReplayFile;
 import com.replaymod.replaystudio.studio.ReplayStudio;
 import com.replaymod.replaystudio.util.I18n;
 import com.replaymod.simplepathing.ReplayModSimplePathing;
-import de.johni0702.minecraft.gui.container.GuiScreen;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.options.Option;
 import net.minecraft.resource.DirectoryResourcePack;
@@ -29,21 +26,12 @@ import net.minecraft.text.Text;
 import net.minecraft.text.TranslatableText;
 import net.minecraft.util.Formatting;
 import net.minecraft.util.Identifier;
-import org.apache.commons.io.FileUtils;
-import org.apache.commons.io.FilenameUtils;
 
 import java.io.ByteArrayInputStream;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
-import java.net.URLDecoder;
 import java.nio.charset.StandardCharsets;
-import java.nio.file.DirectoryStream;
-import java.nio.file.FileVisitResult;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.SimpleFileVisitor;
-import java.nio.file.attribute.BasicFileAttributes;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.ExecutionException;
@@ -73,6 +61,8 @@ public class ReplayMod implements Module, Scheduler {
     private final List<Module> modules = new ArrayList<>();
 
     private final GuiBackgroundProcesses backgroundProcesses = new GuiBackgroundProcesses();
+    public final ReplayFoldersService folders = new ReplayFoldersService(settingsRegistry);
+    public final ReplayFilesService files = new ReplayFilesService(folders);
 
     /**
      * Whether the current MC version is supported by the embedded ReplayStudio version.
@@ -114,58 +104,6 @@ public class ReplayMod implements Module, Scheduler {
 
     public SettingsRegistry getSettingsRegistry() {
         return settingsRegistry;
-    }
-
-    public Path getReplayFolder() throws IOException {
-        String str = getSettingsRegistry().get(Setting.RECORDING_PATH);
-        return Files.createDirectories(getMinecraft().runDirectory.toPath().resolve(str));
-    }
-
-    /**
-     * Folder into which replay backups are saved before the MarkerProcessor is unleashed.
-     */
-    public Path getRawReplayFolder() throws IOException {
-        return Files.createDirectories(getReplayFolder().resolve("raw"));
-    }
-
-    /**
-     * Folder into which replays are recorded.
-     * Distinct from the main folder, so they cannot be opened while they are still saving.
-     */
-    public Path getRecordingFolder() throws IOException {
-        return Files.createDirectories(getReplayFolder().resolve("recording"));
-    }
-
-    /**
-     * Folder in which replay cache files are stored.
-     * Distinct from the recording folder cause people kept confusing them with recordings.
-     */
-    public Path getCacheFolder() throws IOException {
-        String str = getSettingsRegistry().get(Setting.CACHE_PATH);
-        Path path = getMinecraft().runDirectory.toPath().resolve(str);
-        Files.createDirectories(path);
-        try {
-            Files.setAttribute(path, "dos:hidden", true);
-        } catch (UnsupportedOperationException ignored) {
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        return path;
-    }
-
-    private static final PercentEscaper CACHE_FILE_NAME_ENCODER = new PercentEscaper("-_ ", false);
-
-    public Path getCachePathForReplay(Path replay) throws IOException {
-        Path replayFolder = getReplayFolder();
-        Path cacheFolder = getCacheFolder();
-        Path relative = replayFolder.toAbsolutePath().relativize(replay.toAbsolutePath());
-        return cacheFolder.resolve(CACHE_FILE_NAME_ENCODER.escape(relative.toString()));
-    }
-
-    public Path getReplayPathForCache(Path cache) throws IOException {
-        String relative = URLDecoder.decode(cache.getFileName().toString(), "UTF-8");
-        Path replayFolder = getReplayFolder();
-        return replayFolder.resolve(relative);
     }
 
     public static final DirectoryResourcePack jGuiResourcePack = createJGuiResourcePack();
@@ -230,114 +168,7 @@ public class ReplayMod implements Module, Scheduler {
         }
         //#endif
 
-        runPostStartup(() -> {
-            final long DAYS = 24 * 60 * 60 * 1000;
-
-            // Cleanup any cache folders still remaining in the recording folder (we once used to put them there)
-            try {
-                Files.walkFileTree(getReplayFolder(), new SimpleFileVisitor<Path>() {
-                    @Override
-                    public FileVisitResult preVisitDirectory(Path dir, BasicFileAttributes attrs) throws IOException {
-                        String name = dir.getFileName().toString();
-                        if (name.endsWith(".mcpr.cache")) {
-                            FileUtils.deleteDirectory(dir.toFile());
-                            return FileVisitResult.SKIP_SUBTREE;
-                        }
-                        return super.preVisitDirectory(dir, attrs);
-                    }
-                });
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-
-            // Cleanup raw folder content three weeks after creation (these are pretty valuable for debugging)
-            try (DirectoryStream<Path> paths = Files.newDirectoryStream(getRawReplayFolder())) {
-                for (Path path : paths) {
-                    if (Files.getLastModifiedTime(path).toMillis() + 21 * DAYS < System.currentTimeMillis()) {
-                        Files.delete(path);
-                    }
-                }
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-
-            // Move anything which is still in the recording folder into the regular replay folder
-            // so it can be opened and/or recovered
-            try (DirectoryStream<Path> paths = Files.newDirectoryStream(getRecordingFolder())) {
-                for (Path path : paths) {
-                    Path destination = getReplayFolder().resolve(path.getFileName());
-                    if (Files.exists(destination)) {
-                        continue; // better play it save
-                    }
-                    Files.move(path, destination);
-                }
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-
-            // Cleanup cache folders 7 days after last modification or when its replay is gone
-            try (DirectoryStream<Path> paths = Files.newDirectoryStream(getCacheFolder())) {
-                for (Path path : paths) {
-                    if (Files.isDirectory(path)) {
-                        Path replay = getReplayPathForCache(path);
-                        long lastModified = Files.getLastModifiedTime(path).toMillis();
-                        if (lastModified + 7 * DAYS < System.currentTimeMillis() || !Files.exists(replay)) {
-                            FileUtils.deleteDirectory(path.toFile());
-                        }
-                    }
-                }
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-
-            // Cleanup deleted corrupted replays
-            try (DirectoryStream<Path> paths = Files.newDirectoryStream(getReplayFolder())) {
-                for (Path path : paths) {
-                    String name = path.getFileName().toString();
-                    if (name.endsWith(".mcpr.del") && Files.isDirectory(path)) {
-                        long lastModified = Files.getLastModifiedTime(path).toMillis();
-                        if (lastModified + 2 * DAYS < System.currentTimeMillis()) {
-                            FileUtils.deleteDirectory(path.toFile());
-                        }
-                    }
-                }
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-
-            // Restore corrupted replays
-            try (DirectoryStream<Path> paths = Files.newDirectoryStream(getReplayFolder())) {
-                for (Path path : paths) {
-                    String name = path.getFileName().toString();
-                    if (name.endsWith(".mcpr.tmp") && Files.isDirectory(path)) {
-                        Path original = path.resolveSibling(FilenameUtils.getBaseName(name));
-                        Path noRecoverMarker = original.resolveSibling(original.getFileName() + ".no_recover");
-                        if (Files.exists(noRecoverMarker)) {
-                            // This file, when its markers are processed, doesn't actually result in any replays.
-                            // So we don't really need to recover it either, let's just get rid of it.
-                            FileUtils.deleteDirectory(path.toFile());
-                            Files.delete(noRecoverMarker);
-                            continue;
-                        }
-                        new RestoreReplayGui(this, GuiScreen.wrap(mc.currentScreen), original.toFile()).display();
-                    }
-                }
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-
-            // Cleanup leftover no_recover files
-            try (DirectoryStream<Path> paths = Files.newDirectoryStream(getReplayFolder())) {
-                for (Path path : paths) {
-                    String name = path.getFileName().toString();
-                    if (name.endsWith(".no_recover")) {
-                        Files.delete(path);
-                    }
-                }
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        });
+        runPostStartup(() -> files.initialScan(this));
     }
 
     @Override
@@ -436,18 +267,5 @@ public class ReplayMod implements Module, Scheduler {
         } else {
             return new ReplayStudio().isCompatible(fileFormatVersion, protocolVersion, MCVer.getProtocolVersion());
         }
-    }
-
-    public ReplayFile openReplay(Path path) throws IOException {
-        return openReplay(path, path);
-    }
-
-    public ReplayFile openReplay(Path input, Path output) throws IOException {
-        return new ZipReplayFile(
-                new ReplayStudio(),
-                input != null ? input.toFile() : null,
-                output.toFile(),
-                getCachePathForReplay(output).toFile()
-        );
     }
 }
